requirements.txt
asttokens==2.4.1
blinker==1.7.0
click==8.1.7
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
Flask==3.0.2
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.5.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
platformdirs==4.1.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pyzmq==25.1.2
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
wcwidth==0.2.12
Werkzeug==3.0.1
